move JourneyTile to domain/journey/common
